Removing references to WMM.
