+++ eoxmagmod/eoxmagmod/Makefile
@@ -22,29 +22,29 @@ CCFLAGS= -c -O2 -Wall -Wno-unused-function -fPIC $(INCS)
 #CCFLAGS= -c -O2 -g -DDEBUG=1 -Wall -fPIC $(INCS)
 #CCFLAGS= -c -O2 -Wall -fPIC $(INC)
 
-HDR_WMM=pywmm_aux.h pywmm_sheval.h pywmm_sphargrd.h pywmm_spharpot.h\
-    pywmm_legendre.h pywmm_lonsincos.h pywmm_relradpow.h\
-    pywmm_vrot_common.h pywmm_vrot_sph2cart.h pywmm_vrot_sph2geod.h\
-    pywmm_vrot_cart2sph.h pywmm_cconv.h pywmm_coord.h pywmm_exc.h
-OBJ_WMM=pywmm.o
-DST_WMM=_pywmm.so
+HDR_WMM=pymm_aux.h pymm_sheval.h pymm_sphargrd.h pymm_spharpot.h\
+    pymm_legendre.h pymm_lonsincos.h pymm_relradpow.h\
+    pymm_vrot_common.h pymm_vrot_sph2cart.h pymm_vrot_sph2geod.h\
+    pymm_vrot_cart2sph.h pymm_cconv.h pymm_coord.h pymm_exc.h
+OBJ_WMM=pymm.o
+DST_WMM=_pymm.so
 
-HDR_QD=pywmm_aux.h pyqd_exc.h\
+HDR_QD=pymm_aux.h pyqd_exc.h\
     pyqd_eval_qdlatlon.h pyqd_eval_mlt.h pyqd_eval_subsol.h pyqd_eval_apex.h
 OBJ_QD=pyqd.o
 DST_QD=_pyqd.so
 
-HDR_SP=pywmm_aux.h pysunpos.h pysunpos_original.h
+HDR_SP=pymm_aux.h pysunpos.h pysunpos_original.h
 OBJ_SP=pysunpos.o
 DST_SP=_pysunpos.so
 
-HDR_TC=pywmm_aux.h pytimeconv_mjd2000_to_decimal_year.h pytimeconv_decimal_year_to_mjd2000.h
+HDR_TC=pymm_aux.h pytimeconv_mjd2000_to_decimal_year.h pytimeconv_decimal_year_to_mjd2000.h
 OBJ_TC=pytimeconv.o
 DST_TC=_pytimeconv.so
 
 # --------------------------------------------
 
-all: pywmm pyqd pysunpos pytimeconv
+all: pymm pyqd pysunpos pytimeconv
 
 %.o: %.c
 	$(CC) $< $(CCFLAGS) -o $@
@@ -52,7 +52,7 @@ all: pywmm pyqd pysunpos pytimeconv
 %.o: %.cpp
 	$(CC) $< $(CCFLAGS) -o $@
 
-pywmm: $(OBJ_WMM)
+pymm: $(OBJ_WMM)
 	$(LD) -shared -o $(DST_WMM) $(OBJ_WMM) $(LIBS_WMM)
 
 pyqd: $(OBJ_QD)
@@ -70,7 +70,7 @@ clean:
 purge: clean
 	$(RM) $(DST_WMM)
 
-pywmm.o: pywmm.c $(HDR_WMM)
+pymm.o: pymm.c $(HDR_WMM)
 pyqd.o: pyqd.c $(HDR_QD)
 pysunpos.o: pysunpos.c $(HDR_SP)
 pytimeconv.o: pytimeconv.c $(HDR_TC)
