+++ eoxmagmod/eoxmagmod/pymm_relradpow.h
@@ -28,11 +28,11 @@
  *-----------------------------------------------------------------------------
 */
 
-#ifndef PYWMM_RELRADPOW_H
-#define PYWMM_RELRADPOW_H
+#ifndef PYMM_RELRADPOW_H
+#define PYMM_RELRADPOW_H
 
 #include "shc.h"
-#include "pywmm_aux.h"
+#include "pymm_aux.h"
 
 /* Earth radius in km */
 #define RADIUS  6371.2
@@ -119,4 +119,4 @@ static PyObject* relradpow(PyObject *self, PyObject *args, PyObject *kwdict)
     return retval;
 }
 
-#endif  /* PYWMM_RELRADPOW_H */
+#endif  /* PYMM_RELRADPOW_H */
