+++ eoxmagmod/eoxmagmod/pyqd_eval_apex.h
@@ -33,7 +33,7 @@
 #ifndef PYQD_EVAL_APEX_H
 #define PYQD_EVAL_APEX_H
 
-#include "pywmm_aux.h"
+#include "pymm_aux.h"
 #include "qdipole/cqdipole.h"
 
 /* python function definition */
