+++ eoxmagmod/eoxmagmod/pyqd_eval_mlt.h
@@ -31,7 +31,7 @@
 #ifndef PYQD_EVAL_MLT_H
 #define PYQD_EVAL_MLT_H
 
-#include "pywmm_aux.h"
+#include "pymm_aux.h"
 #include "qdipole/cqdipole.h"
 
 /* python function definition */
