+++ eoxmagmod/eoxmagmod/pysunpos_original.h
@@ -30,7 +30,7 @@
 #ifndef PYSUNPOS_ORIGINAL_H
 #define PYSUNPOS_ORIGINAL_H
 
-#include "pywmm_aux.h"
+#include "pymm_aux.h"
 #include "sunpos.h"
 #include <math.h>
