+++ eoxmagmod/eoxmagmod/pytimeconv_mjd2000_to_year_fraction.h
@@ -30,7 +30,7 @@
 #ifndef PYTIMECONV_MJD2000_TO_YEAR_FRACTION_H
 #define PYTIMECONV_MJD2000_TO_YEAR_FRACTION_H
 
-#include "pywmm_aux.h"
+#include "pymm_aux.h"
 #include "time_conversion.h"
 #include <math.h>
