+++ eoxmagmod/eoxmagmod/tests/pymm_legendre.py
@@ -32,7 +32,7 @@ from math import pi, sin, sqrt
 from numpy import array
 from numpy.testing import assert_allclose
 from scipy.special import legendre as legendre_polynomial
-from eoxmagmod._pywmm import legendre
+from eoxmagmod._pymm import legendre
 
 
 class TestLegendreFunctions(TestCase):
